created testing utils for generating dummy data, model, and trainer
